Updated Upgrade System
Finished implementing upgrade system including logic that increments the upgrade and deducts the resources from the player.

Added accessor for currentScreen in View. Added non-const accessors to Player, City, and Model classes for direct modification by the controller. Added event logic in controller and new Enums to support this logic in the GameEvent class.